Updates per comments
